pyproject.toml
[project]
name = "projectiq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "chromadb>=1.4.0",
    "extract-msg>=0.55.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "libmagic>=1.0",
    "lxml>=6.0.2",
    "pillow>=12.1.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tesseract>=0.1.3",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.18.26",
]
